pyproject.toml
[tool.poetry]
name = "statsbook"
version = "0.1.0"
description = ""
authors = ["Chansoo Song"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
seaborn = "^0.11.2"
numpy = "^1.22.3"
pandas = "^1.4.2"
matplotlib = "^3.5.1"
scipy = "^1.8.0"
scikit-learn = "^1.0.2"
statsmodels = "^0.13.2"
tensorflow-cpu = "^2.8.0"
Faker = "^13.4.0"
tqdm = "^4.64.0"
jupyter-book = "0.15.1"
jupytext = "^1.13.8"
ipykernel = "^6.13.0"
ghp-import = "^2.0.2"
pydotplus = "^2.0.2"
xgboost = "^1.6.0"
plotly = "^5.7.0"
ipympl = "^0.9.1"
tabulate = "^0.8.9"
mkdocs = "^1.4.2"
mkdocs-material = ">=8.0.0,<9.0.0"
mkdocs-simple-blog = "^0.0.5"
mkdocstrings = "^0.20.0"
mkdocs-jupyter = "^0.22.0"
pymdown-extensions = "^9.9.2"
mkdocs-exclude = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
